% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot.R
\name{myplot}
\alias{myplot}
\title{Add myplot function}
\usage{
myplot(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A plot of x and y.
}
\description{
Add myplot function
}
\examples{
set.seed(312)
x <- rnorm(100)
y <- rnorm(100)
myplot(x, y)
}
